docs : 파일 리팩토링으로 인한 폴더 구조 변경
