--- src/@components/@common/Error404/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Error404 = () => {
-  return <div>404페이지입니다.</div>;
-};
-
-export default Error404;
