--- src/@components/@common/LearningBtnList/style.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const St = {
-  LearningBtnNav: styled.nav`
-    display: flex;
-    justify-content: flex-start;
-    gap: 1rem;
-
-    padding: 1rem 0rem;
-  `,
-};
