+++ src/@components/@common/QuestionTitle/HashTag.tsx
@@ -1,6 +1,22 @@
+import React from 'react';
+import { HashTagList } from '../../../type/problem';
 import styled from 'styled-components';
 
-export const St = {
+const HashTag = (props: HashTagList) => {
+  const { hashTagList } = props;
+
+  return (
+    <St.HashTagWrapper>
+      {hashTagList.map((hashTag) => (
+        <li key={hashTag}>{hashTag}</li>
+      ))}
+    </St.HashTagWrapper>
+  );
+};
+
+export default HashTag;
+
+const St = {
   HashTagWrapper: styled.ul`
     display: flex;
     justify-content: flex-start;
