+++ src/@components/@common/QuestionTitle/index.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import HashTag from './HashTag';
-import { St } from './style';
+import styled from 'styled-components';
 
 interface QuestionTitleProps {
   question: string;
@@ -18,3 +17,19 @@ const QuestionTitle = (props: QuestionTitleProps) => {
 };
 
 export default QuestionTitle;
+
+const St = {
+  QuestionHeader: styled.header`
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    gap: 2rem;
+
+    padding: 1.5rem 0.9rem;
+
+    background-color: ${({ theme }) => theme.colors.Peer_Color_Sky_3};
+  `,
+  Title: styled.h1`
+    ${({ theme }) => theme.fonts.Peer_Noto_B_Title_1};
+  `,
+};
