+++ src/@components/AnswerListPage/index.ts
@@ -1,0 +1,2 @@
+export { default as AnswerCard } from './AnswerCard';
+export { default as DetailModal } from './DetailModal';
