--- src/@components/MainPage/LearningExperience/style.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const St = {
-  LearningExperienceSection: styled.section`
-    display: flex;
-    flex-direction: column;
-    row-gap: 5.5rem;
-
-    width: 100%;
-    margin-bottom: 5.4rem;
-  `,
-};
