--- src/@components/MainPage/PeerMatchingBtn/style.ts
@@ -1,26 +1,0 @@
-import styled from 'styled-components';
-
-export const St = {
-  MatchigBtnWrapper: styled.div`
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-
-    width: 100%;
-    height: 7.2rem;
-    padding: 1.9rem 3.6rem;
-
-    background-color: ${({ theme }) => theme.colors.Peer_Color_Blue};
-    border-radius: 3rem;
-
-    & > span {
-      font-size: 5rem;
-      color: ${({ theme }) => theme.colors.Peer_Color_White_2};
-    }
-  `,
-  MatchingBtn: styled.button`
-    ${({ theme }) => theme.fonts.Peer_Noto_B_Title_2}
-    background: none;
-    border: none;
-  `,
-};
