+++ src/@components/MainPage/index.ts
@@ -1,0 +1,5 @@
+export { default as LearningExperience } from './LearningExperience';
+export { default as LearningRecord } from './LearningRecord';
+export { default as LearningSelect } from './LearningSelect';
+export { default as PeerMatchingBtn } from './PeerMatchingBtn';
+export { default as RecordCard } from './RecordCard';
