+++ src/pages/Callback.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { getUserInfo } from '../../../lib/api/auth';
+import { getUserInfo } from '../lib/api/auth';
 
 const Callback = () => {
   const code = new URL(window.location.href).searchParams.get('login');
@@ -13,8 +13,9 @@ const Callback = () => {
 
   const checkLogin = async () => {
     if (code === 'success') {
-      // const data = await getUserInfo();
-      navigate('/main');
+      const data = await getUserInfo();
+      console.log(data);
+      // navigate('/main');
     } else navigate('/');
   };
   return null;
